% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_guide.R
\name{function_guide}
\alias{function_guide}
\title{Guide for adding function to package}
\usage{
function_guide()
}
\value{
steps
}
\description{
Guide for adding function to package
}
\examples{
function_guide()
}
