% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_mean.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Calculate harmonic mean}
\usage{
harmonic_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{remove NA}
}
\value{
single value
}
\description{
The harmonic mean is calculated as the number of values N
divided by the sum of the reciprocal of the values (1 over each value).
}
\examples{
\dontrun{
harmonic_mean(mtcars[['mpg']])
}
}
